requirements.txt
pillow
numpy
matplotlib
scipy
scikit-learn
torch
torchvision
daltonize
